+++ crates/polkavm-derive-impl-macro/README.md
@@ -1,0 +1,1 @@
+This is an internal crate of PolkaVM; **do not** use it directly!
